backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
structlog>=24.1.0
prometheus-client>=0.20.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
mecademicpy==2.3.1
pythonnet>=3.0.3
pypylon>=3.0.1
pyserial>=3.5
pylablib>=1.4.2
numpy>=1.26.0
opencv-python-headless>=4.9.0.80
